packages/middlewares/embedding-processors/clip-image-processor/src/container/requirements.txt
boto3==1.34.22
ftfy==6.1.3
regex==2023.12.25
tqdm==4.66.1
torchvision==0.16.2
git+https://github.com/openai/CLIP.git
